refactor : result패키지 추가
- LottoRank 클래스 이동
- WinningLotto 클래스 이동
- WinningResult 클래스 이동
