+++ src/main/java/lotto/domain/result/WinningResult.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.result;
 
 import java.util.Map;
 import java.util.stream.Stream;
